simple-knowledge-qa/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
langchain-community==0.0.20
sentence-transformers==2.3.1
faiss-cpu==1.7.4
numpy==1.24.3
google-generativeai==0.8.3
python-dotenv==1.0.0
